% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_functions.r
\name{set_splicr_conflict_handling}
\alias{set_splicr_conflict_handling}
\title{Change how multi-pipe (\%!>\%) handles argument conflicts}
\usage{
set_splicr_conflict_handling(cond = c("warning", "error", "message",
  "silent"))
}
\arguments{
\item{cond}{The condition triggered by conflicting arguments}
}
\description{
By default, attempting to use the same argument on the right
left sides of a multi-pipe (e.g., \code{list(a = 1) \%!>\% fun(a = 2)} )
produces a warning, and the right-hand side argument is used.
This sets a global option that changes the warning into
an error, message, or silent condition.
}
\seealso{
\%!>\%
}
